Split off host name fetching to helper class in preparation of future commands
